signals: route hx_printf* through a chat-log-line signal
Post-Phase-3 follow-up. The vtable migration in Phase 3 closed
every model→view edge except one: hx_printf and hx_printf_prefix.
Model files (rcv.c, network.c, commands.c, banner.c, tracker.c,
toolbar.c, xfers.c) call these by name to write info lines into
chat output; they used to live in chat.c and reach the xtext
widget directly. After this commit they emit through GtkhxSession
like every other model→view notification.

What lands:

  - src/gtkhx_log.{c,h}: new home for hx_printf / hx_printf_prefix.
    Pure glib + stdio — no GTK. They share a static log_line_emit
    helper that vsnprintf's into a heap-or-stack buffer and emits
    the new 'chat-log-line' signal (htlc, cid, body).

  - src/gtkhx_session.{c,h}: 'chat-log-line' signal added. Same
    emit-helper pattern as the other signals.

  - src/chat.c: hx_printf / hx_printf_prefix gone; replaced with
    chat_log_line_handler, a small static-ish handler that fans
    the body through the existing xoutput_chat. Connected once in
    gtkhx_connect_signals at startup.

  - src/chat.h: the hx_printf decls now come via #include
    "gtkhx_log.h" which chat.h forwards — every existing caller
    that #include "chat.h" still gets the signatures, so no
    call-site churn.

  - meson.build: src/gtkhx_log.c added to the sources list.

The new G_GNUC_PRINTF attributes on the helpers caught five
pre-existing format-string bugs the codebase has been silently
carrying. Fixed inline:

  - commands.c command-not-found: ptrdiff_t %.* arg → (int) cast.
  - network.c GZIP deflate / inflate counters were guint32 printed
    with %u, but the field types are actually unsigned long. Cast
    to (unsigned long) and use %lu.
  - network.c htlc_read / htlc_write error log printed ssize_t with
    %d. Use %zd.

24/24 unit + proto tests pass. The chat-log-line signal closes the
last unsignaled model→view edge; model files no longer reach into
view-side code by name to log.

+++ src/commands.c
@@ -618,5 +618,5 @@ void hx_command (char *str, guint32 cid)
 	} while (cmd <= last_command && cmd->name[0] == *str);
 
   notfound:
-	hx_printf_prefix(&the_session.htlc, cid, INFOPREFIX, "%.*s: command not found\n", p - str, str);
+	hx_printf_prefix(&the_session.htlc, cid, INFOPREFIX, "%.*s: command not found\n", (int)(p - str), str);
 }
